fix: move unit tests in codex-rs/core/src/codex.rs into their own file
